+++ Pandora/Paradigm/Structure/Some.hs
@@ -1,5 +1,6 @@
 module Pandora.Paradigm.Structure.Some (module Exports) where
 
+import Pandora.Paradigm.Structure.Some.Vector as Exports
 import Pandora.Paradigm.Structure.Some.Rose as Exports
 import Pandora.Paradigm.Structure.Some.Splay as Exports
 import Pandora.Paradigm.Structure.Some.Binary as Exports