+++ Pandora/Paradigm/Structure/Some/Vector.hs
@@ -1,0 +1,5 @@
+module Pandora.Paradigm.Structure.Some.Vector where
+
+import Pandora.Paradigm.Structure.Ability.Monotonic (Monotonic)
+
+data Vector a = forall v . Monotonic a v => Vector v